hwmon: stts751: cache therm flag in stts751_update_alert()
Therm flag in in the same register of alert flags. Those flags clears
on read, so we can't read this register anywhere without care.

So we need to treat therm flags like alarm flags, and read it in
the same manner/place.

This is preparatory for adding a show function for therm state

Signed-off-by: Andrea Merello <andrea.merello@gmail.com>