hwmon: stts751: fix the alarm caching mechanism was ineffective
We need to briefly cache the alarms' flags because when reading the
status register all the flags get cleared until the next conversion.

We set our cached flags when we see a flag in the status reg, and we
clear them when the cache expires (unless they are still asserted in
the status reg).

This requires the cache time to be greater than the actual conversion
time (that is probably not so precise due to RC oscillator).

Note that we can read the status reg at any time for the sake of the
alert interrupt handler.

For this reason, we don't simply reset the cache timestam at every reg
read (because we could reset it too often ending up in never clearing
the cache, that never expires); instead we reset it when we see an
asserted flag.

This makes sure that we lock to the internal converter period: our
cache get cleared after the given time is past since we saw the alarm
asserted.

On the other hand we need a real cache time for reads from userspace
to avoid userspace to potentially starve the I2C bus.

Signed-off-by: Andrea Merello <andrea.merello@gmail.com>